backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision & Video Processing
opencv-python==4.8.1.78
ultralytics==8.0.220
moviepy==1.0.3

# Data Processing
pandas==2.1.3
numpy==1.26.2

# AI Integration
google-generativeai==0.3.1

# Database
psycopg2-binary==2.9.9
supabase==2.3.0  # Optional, for legacy compatibility
postgrest==0.13.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
